Check the influence of the size of the neighbourhood

+++ scripts/dsoRun.py
@@ -68,34 +68,50 @@ sequences = [
 
 runsPerSequence = 1;
 
-#mainDatasetPath = '/mnt/data/Datasets/dso';
-#mainDatasetPath = '/media/michalnowicki/MNowicki-Private/DSO/sequences'
-mainDatasetPath='/home/michal/dsoDataset'
+if len(sys.argv) != 2:
+	print('Please provide the name of the host machine, e.g. python orbslamRun.py lrm2')
+else:
+
+	mainDatasetPath = '';
+
+	# PPCM
+	if "ppcm" in sys.argv[1]:
+		mainDatasetPath = '/home/michal/dsoDataset';
+
+	# LRM2
+	if "lrm2" in sys.argv[1]:
+		mainDatasetPath = '/mnt/data/Datasets/dso';
+
+	# Toshiba Portege Z30
+	if "local" in sys.argv[1]:
+		mainDatasetPath = '/media/michalnowicki/MNowicki-Private/DSO/sequences';
+
+	print 'mainDatasetPath: ' + mainDatasetPath
 
 
-detectionTypes = [6, 6, 6, 6, 6, 6] #[4, 4]#[] 1, 1, 1];
-detectionTypeFastThreshold = [0] #[7, 2];#[0, 15, 10, 5];
-harrisK = [0.002, 0.005, 0.01, 0.02, 0.04, 0.08];
+	detectionTypes = [6] #[6, 6, 6, 6, 6, 6] #[4, 4]#[] 1, 1, 1];
+	detectionTypeFastThreshold = [0] #[7, 2];#[0, 15, 10, 5];
+	harrisK = [0.01];#[0.002, 0.005, 0.01, 0.02, 0.04, 0.08];
 
-# Clear the results directories
-for (det, hark) in zip(detectionTypes, harrisK):
-	if not os.path.exists("results/det_"+str(det)+"_"+str(hark)):
-		os.makedirs("results/det_"+str(det)+"_"+str(hark));
-	else:
-		call('rm results/det_'+str(det)+'_'+str(hark) +'/*', shell=True);
-
-for seq in sequences:
-
+	# Clear the results directories
 	for (det, hark) in zip(detectionTypes, harrisK):
-		
-		for runId in range(0, runsPerSequence):
-			print("Current sequence: " + seq);
+		if not os.path.exists("results/det_"+str(det)+"_"+str(hark)):
+			os.makedirs("results/det_"+str(det)+"_"+str(hark));
+		else:
+			call('rm results/det_'+str(det)+'_'+str(hark) +'/*', shell=True);
 
-			print('./dso_dataset files='+mainDatasetPath+'/sequence_'+ seq +'/images.zip calib='+mainDatasetPath+'/sequence_'+ seq +'/camera.txt gamma='+mainDatasetPath+'/sequence_'+ seq +'/pcalib.txt vignette='+mainDatasetPath+'/sequence_'+ seq +'/vignette.png preset=0 mode=0 nogui=1 reverse=0 quiet=1 detectionType=' + str(det) + ' harrisK=' + str(hark));
+	for seq in sequences:
 
-			# Copy to currently used settings
-			call('./dso_dataset files='+mainDatasetPath+'/sequence_'+ seq +'/images.zip calib='+mainDatasetPath+'/sequence_'+ seq +'/camera.txt gamma='+mainDatasetPath+'/sequence_'+ seq +'/pcalib.txt vignette='+mainDatasetPath+'/sequence_'+ seq +'/vignette.png preset=0 mode=0 nogui=1 reverse=0 quiet=1 detectionType=' + str(det) + ' harrisK=' + str(hark), shell=True);
+		for (det, hark) in zip(detectionTypes, harrisK):
 
-			# Run software
-			call('mv result.txt results/det_'+str(det)+'_'+str(hark) +'/sequence_' + str(seq) + '_' + str(runId) + '.txt', shell=True);
+			for runId in range(0, runsPerSequence):
+				print("Current sequence: " + seq);
+
+				print('./dso_dataset files='+mainDatasetPath+'/sequence_'+ seq +'/images.zip calib='+mainDatasetPath+'/sequence_'+ seq +'/camera.txt gamma='+mainDatasetPath+'/sequence_'+ seq +'/pcalib.txt vignette='+mainDatasetPath+'/sequence_'+ seq +'/vignette.png preset=0 mode=0 nogui=1 reverse=0 quiet=1 detectionType=' + str(det) + ' harrisK=' + str(hark));
+
+				# Copy to currently used settings
+				call('./dso_dataset files='+mainDatasetPath+'/sequence_'+ seq +'/images.zip calib='+mainDatasetPath+'/sequence_'+ seq +'/camera.txt gamma='+mainDatasetPath+'/sequence_'+ seq +'/pcalib.txt vignette='+mainDatasetPath+'/sequence_'+ seq +'/vignette.png preset=0 mode=0 nogui=1 reverse=0 quiet=1 detectionType=' + str(det) + ' harrisK=' + str(hark), shell=True);
+
+				# Run software
+				call('mv result.txt results/det_'+str(det)+'_'+str(hark) +'/sequence_' + str(seq) + '_' + str(runId) + '.txt', shell=True);
